Attempted Fix at everything
Now works in RSBot Env. (sort of). Go to Varrock (Lumbridge is incorrect
location) and Walk through Varrock for a bit and then use the GUI to
have it walk back to the main square in Varrock (this sort of works most
of the time). Still not smooth, but progress...