Implemented a StepTask class
A Steptask has a step() method that executes one step. this method is
called contiuously while the remaining amoun of steps > 0. This makes is
easier to implement tasks that clearly consist of multiple steps. It is
also similar to the way ActiveScript works, so that other script writers
can easily plug into our plugin with their existing scripts.