Fully working pathfinding and travelTo command!!
Fixed and tweaked the hell out of the collision detection, and finally
got it to work (or so it seems). Appearantly, Java reflect CAN be used,
as long as you stay away from any RSBot classes. Using some shady tricks
(basically everthing in the reflect package is shady), I was able to
make it easily extendible, too. Now for collecting raw values of walls,
doors, etc...