Implemented the MapViewer
It allows you to view the contents of the world map in a graphical
representation. Just run the main of MapViewer to see. refer to the
JavaDoc or the comments inside the paint method for the Color Mapping
that was used. Alter the view by adjusting the Rectangle and scale
parameters inside the MapViewer class to view different portions of the
World Map. (It tends to get slow when trying to view 3000 x 3000 tiles
at once). The current view is Varrock and surroundings.