Added various new Classes and improvements
Added a DefaultInteractor that simply clicks the object. This prevents many separate Interactors.
Added a Tilefactory which will act as the main class for instantiating GameTile subclasses.
Modified Interactor so that the isMoreFavorableThan method checks the compatible classes by default (was abstract)
Added a cache for the Widgets requested through the RSInteractor.