Removed Mac OS-specific code from project
Because we're not planning to support it in the near future, and the existing measures didn't work anyway.
Also cleaned unsused files and libraries from the repository, most notably the jace jar files, that were
only capable of generating proxies. They are removed because we have no need for them.