Changed UpdaterRunner to use the Files and Path classes.
Also minimized the OptionList class, as it contained many unused classes
(and missed a class we actually needed: a JavaAgent Option).
Changed exception handling, and added native method call in case the updater
data is not found, allowing the native client to disable features based on
the presence of updater data.