Added more specific documentation to ExpressionParser.
The evaluate function now is supposed to return a boxed primitive type when the result is a primitive,
eliminating the problem with accessing primitive data types through the ExpressionParser. The classes
that box each of the primitive types in Java need to be included as proxy classes to make this work.