Included a change in the updater XML format.
long fields can now also have modifiers, indicated by an XML tag
'longmodifier' (the same way as the 'modifier' tag). These modifiers
are now also included in the modifier map.

This change still needs to be tested, as there is currently no updater
data available to test with. Tests will be conducted as soon as the data
is available.